refactor: rename manager dispatcher
